requirement.txt
Django==1.9.5
MySQL-python==1.2.5
django-registration==2.0.4
argparse==1.2.1
wsgiref==0.1.2
